+++ issurge/github.py
@@ -36,10 +36,21 @@ def available_issue_types():
     response = json.loads(call_api("GET", f"/orgs/{repo.owner}/issue-types") or "[]")
     return [t["name"] for t in response]
 
+def issue_id(number: int):
+    issue_id = call_repo_api("GET", f"issues/{number}", jq=".id")
+    if not issue_id:
+        raise Exception(f"Could not retrieve issue ID for issue #{number}")
+    return int(issue_id)
+
+
 type HTTPMethod = Literal["GET", "POST", "PATCH", "DELETE"]
 
+
 def call_api(
-    method: Literal["GET", "POST", "PATCH", "DELETE"], route: str, **body_fields: Any
+    method: Literal["GET", "POST", "PATCH", "DELETE"],
+    route: str,
+    jq="",
+    **body_fields: Any,
 ):
     cmd = ["gh", "api"]
     if method != "GET":
@@ -53,10 +64,19 @@ def call_api(
         else:
             cmd += ["-F", f"{key}={json.dumps(value)}"]
 
+    if jq:
+        cmd += ["--jq", jq]
+
     return run(cmd)
 
+
 def call_repo_api(
-    method: Literal["GET", "POST", "PATCH", "DELETE"], route: str, **body_fields: Any
+    method: Literal["GET", "POST", "PATCH", "DELETE"],
+    route: str,
+    jq="",
+    **body_fields: Any,
 ):
     repo = repo_info()
-    return call_api(method, f"/repos/{repo.owner}/{repo.repo}/{route}", **body_fields)
+    return call_api(
+        method, f"/repos/{repo.owner}/{repo.repo}/{route}", jq=jq, **body_fields
+    )