+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2026-02-11
+
 ## [1.3.2] - 2026-02-09
 
 ### Added
@@ -152,7 +154,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/gwennlbh/issurge/compare/v1.3.2...HEAD
+[Unreleased]: https://github.com/gwennlbh/issurge/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/gwennlbh/issurge/compare/v1.3.2...v1.4.0
 [1.3.2]: https://github.com/gwennlbh/issurge/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/gwennlbh/issurge/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/gwennlbh/issurge/compare/v1.2.1...v1.3.0